Added range option
You can now specify a range of questions with which to create the test.
You do this using the -r or --range flags.  I had to change a few other
tests that I will need to clean up later, but I included them in this
commit out of because I'm currently a bit too confused to pick them
apart.