Quick fix for bug when no tests are present.
When no tests are in the test dir, the program was throwing an error
because of a line that I have in the lib/study_buddy (untested) portion
of my code.  I solved it by just checking the command given. If it's
"make" then it should work correctly.  However, this is broken I'm sure
and needs to be fixed eventually.